# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CreateWorkTreeRun {
  /** @minLength 1 */
  goal: string;
  model?: string;
}

export interface WorkTreeRun {
  id: number;
  goal: string;
  status: string;
  model: string;
  report: string;
  error: string;
  createdAt: string;
  updatedAt: string;
}

export interface WorkTreeRunList {
  runs: WorkTreeRun[];
}

export interface WorkTreeNode {
  id: number;
  runId: number;
  parentId: number | null;
  title: string;
  detail: string;
  kind: string;
  status: string;
  depth: number;
  position: number;
  result: string;
  verification: string;
  attempts: number;
  createdAt: string;
  updatedAt: string;
}

export interface WorkTreeRunDetail {
  run: WorkTreeRun;
  nodes: WorkTreeNode[];
}