# lib/api-zod/src/generated/types/createWorkTreeRun.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWorkTreeRun {
  /** @minLength 1 */
  goal: string;
  model?: string;
}