# lib/api-zod/src/generated/types/workTreeRun.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkTreeRun {
  id: number;
  goal: string;
  status: string;
  model: string;
  report: string;
  error: string;
  createdAt: string;
  updatedAt: string;
}