# lib/api-zod/src/generated/types/workTreeRunDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkTreeNode } from './workTreeNode';
import type { WorkTreeRun } from './workTreeRun';

export interface WorkTreeRunDetail {
  run: WorkTreeRun;
  nodes: WorkTreeNode[];
}