# lib/api-zod/src/generated/types/workTreeRunList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkTreeRun } from './workTreeRun';

export interface WorkTreeRunList {
  runs: WorkTreeRun[];
}